WebRTC 로그인 요청을 수신하면 SIP REGISTER 를 전송하여서 실제 SIP 서버에 로그인하는 기능을 추가함

+++ SipPlatform/StringUtility.cpp
@@ -450,7 +450,7 @@ bool IsPrintString( const char * pszText, int iTextLen )
 {
 	for( int i = 0; i < iTextLen; ++i )
 	{
-		if( isprint( pszText[i] ) == 0 ) return false;
+		if( isprint( (uint8_t)pszText[i] ) == 0 ) return false;
 	}
 
 	return true;
